Starting to visualize it
- All visuals have their compile errors sorted out*
- I realize that flywheel flipped the semantics of rotate so that it now
  takes radians instead of degrees. I fixed everything I noticed but
  this will need follow up.
- Implement collectCrumblingInstances for every visual.

* Flywheel needs to re-implement stealInstance

+++ src/main/java/com/simibubi/create/foundation/ponder/element/WorldSectionElement.java
@@ -239,16 +239,16 @@ public class WorldSectionElement extends AnimatedSceneElement {
 			double rotY = Mth.lerp(pt, prevAnimatedRotation.y, animatedRotation.y);
 			TransformStack.of(ms)
 				.translate(centerOfRotation)
-				.rotateX(rotX)
-				.rotateZ(rotZ)
-				.rotateY(rotY)
+				.rotateX((float) rotX)
+				.rotateZ((float) rotZ)
+				.rotateY((float) rotY)
 				.translateBack(centerOfRotation);
 			if (stabilizationAnchor != null) {
 				TransformStack.of(ms)
 					.translate(stabilizationAnchor)
-					.rotateX(-rotX)
-					.rotateZ(-rotZ)
-					.rotateY(-rotY)
+					.rotateX((float) -rotX)
+					.rotateZ((float) -rotZ)
+					.rotateY((float) -rotY)
 					.translateBack(stabilizationAnchor);
 			}
 		}
